+++ code/src/nuvla/api/resources/__init__.py
@@ -1,0 +1,6 @@
+from .callback import Callback
+from .credential import Credential
+from .deployment import Deployment
+from .deployment_parameter import DeploymentParameter
+from .module import Module
+from .notification import Notification